reverted the changes from 'ISC' to 'BUNDY' in copyright notices.
it's not a good practice to modify an existing copyright notice; if needed,
a new noticed should be added to the existing one.

(besides, it should have not even been 'ISC' in the first place, as queryperf++
is not an ISC project.  this should be corrected, but the fix should better
be done separately).